send email after save event

+++ rosach/static/send_email.js
@@ -21,7 +21,7 @@ $(document).ready(function() {
         }
     });
 
-    $("#send-email-modal .delete-and-send-button").click(function() {
+    function sendEmail(deleteIt) {
         var id = $('#send-email-modal input[name=event-id]').val();
         var form = $('#send-email-modal').find('form');
 
@@ -35,14 +35,27 @@ $(document).ready(function() {
             data: form.serialize(),
             dataType: 'json',
             success: function(data) {
-                deleteDevent(id);
+                if (deleteIt) {
+                    deleteDevent(id);
+                } else {
+                    $("#send-email-modal").modal('hide');
+                }
             },
             error: function(data) {
                 $('#send-email-modal div[role=alert]').html(data.responseJSON['message']);
                 $('#send-email-modal div[role=alert]').show();
             }
         });
+    }
+
+    $("#send-email-modal .send-button").click(function() {
+        sendEmail(false);
     });
+
+    $("#send-email-modal .delete-and-send-button").click(function() {
+        sendEmail(true);
+    });
+
     $('#email-message').summernote({
         height: 300
     });
